Julia set finished
Hello all,
I already finished the code of Julia Set. If you check the code, you will find a new .java file. It is the class of doublePoint. I created it since x and y coordinate always appear at the same time, but the existing class Point only can deal with the type of int. You can use this class in your part and in the future for convenience.