user is now able to change the escape distance in the menu.
Change the variable of _escapeDistance in generateFractal to be public
in order to modify it.

Change the variable of _model in userInterface to be public in order to
modify it.

Add a menu item "input" in menu "file". when user click it, a input
dialog will be pop out and enter. 

work need to do: need to find the range of escape distance, tell the
user about it, and if user enter a value out of the range, throw an
exception or let user enter another value.